send messages through entirely api-driven interface
